Add strict Lightroom check, various bug fixes for overwrite and check
If -s (--strict) argument is given it will check the path too and not
only the file basename.

The following bug fixes have been done:
- if verbose was not given it was unset and so google lookup debug
output crashed. Is set to 0 on unset
- if data is found multiple times in the Lightroom DB we print out a
different warning and not put it into the not found group
- after a google lookup write the data back into the base set for debug
output purposes
- if Lightroom data is used for overwrite data was never written because
the unset check was done on the already set data and not the original
unset data